use long for overall required

+++ module/src/main/java/net/savantly/sprout/franchise/domain/operations/qai/score/QAAScoreDto.java
@@ -15,7 +15,7 @@ public class QAAScoreDto {
 	private String submissionId;
 	private long overallAvailable;
 	private long overallNA;
-	private BigDecimal overallRequired;
+	private long overallRequired;
 	private long overallScore;
 	private BigDecimal overallRating;
